feat: extracting the goals functions into another file, adding the template golang file, adding the migration for the templates
