+++ router/Groups/entry-templates.go
@@ -1,0 +1,8 @@
+package groups
+
+import (
+	"net/http"
+
+	"example.com/m/v2/middlewares"
+	"github.com/gin-gonic/gin"
+)
